changed time to long

+++ Server/src/Game/Time/Time.java
@@ -8,7 +8,7 @@ import Game.Time.DailyLoop.PartOfDay;
     when the Game is started, the current time from Clocks is saved, 
     getTime returns value current time from clocks - the time saved when the Game started */
 public class Time {
-    private int inception;
+    private long inception;
     private Clocks clocks;
     public volatile PartOfDay partOfDay;
     public Time(World game, Clocks clocks, DailyLoop dailyLoop){
@@ -16,7 +16,7 @@ public class Time {
         this.clocks = clocks;
         partOfDay = dailyLoop.partsOfDay[0];
     }
-    public int getTime(){
+    public long getTime(){
         return clocks.time - inception;
     }
 
@@ -24,15 +24,15 @@ public class Time {
      * 1 ticks corresponds to 3 minutes
      * 24x20, 480 ticks corresponds to 1 day
      * 30x24x20, 1440 ticks corrensponds to 1 month*/
-    public int ticksOfToday(){
+    public long ticksOfToday(){
         return getTime() % 1440;
     }
 
-    public int tickOfThisMonth(){
+    public long tickOfThisMonth(){
         return getTime() % 43200; 
     }
 
-    public int ticksOfThisHour(){
+    public long ticksOfThisHour(){
         return getTime() % 60;
     }
 
@@ -40,43 +40,43 @@ public class Time {
         return "Now is " + (getMonth(getTime()) + 1) + "th Month, " + (getDay(tickOfThisMonth()) + 1) + "th Day of this Month, " + (getHours(ticksOfToday()) + 1) + "th hour of this day and " + (getMinutes(ticksOfThisHour()) + 1) + "th minute of this hour." ;
     }
 
-    public static int getMinutes(int numberOfTicks){
+    public static long getMinutes(long numberOfTicks){
         return numberOfTicks;
     }
 
-    public static float getHoursFloat(int numberOfTicks){
+    public static float getHoursFloat(long numberOfTicks){
         return ((float) numberOfTicks) / 60f;
     }
 
-    public static int getHours(int numberOfTicks){
+    public static long getHours(long numberOfTicks){
         return numberOfTicks / 60;
     }
 
-    public static float getDayFloat(int numberOfTicks){
+    public static float getDayFloat(long numberOfTicks){
         return ((float) numberOfTicks) / 1440f;
     }
 
-    public static int getDay(int numberOfTicks){
+    public static long getDay(long numberOfTicks){
         return numberOfTicks / 1440;
     }
 
-    public static float getMonthFloat(int numberOfTicks){
+    public static float getMonthFloat(long numberOfTicks){
         return ((float) numberOfTicks) / 43200f;
     }
 
-    public static int getMonth(int numberOfTicks){
+    public static long getMonth(long numberOfTicks){
         return numberOfTicks / 43200;
     }
 
-    public static int hoursToTicks(int hours){
+    public static long hoursToTicks(long hours){
         return hours * 60;
     }
 
-    public static int daysToTicks(int days){
+    public static long daysToTicks(long days){
         return days * 1440;
     }
 
-    public static int monthsToTicks(int months){
+    public static long monthsToTicks(long months){
         return months * 43200;
     }
 }
